# openapi.json
{
  "components": {
    "schemas": {
...
      "CreateEventRequest": {
        "properties": {
          "amountMinor": {
            "format": "int64",
            "type": "integer"
          },
          "assignedCrewId": {
            "type": "string"
          },
          "durationSeconds": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string"
          },
          "totalTickets": {
            "format": "int64",
            "type": "integer"
          }
        },
        "required": [
          "amountMinor",
          "durationSeconds",
          "name",
          "totalTickets"
        ],
        "type": "object"
      },
...
      "Error": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "example": "invalid_request",
                "type": "string"
              },
              "field": {
                "example": "startsAt",
                "type": "string"
              },
              "message": {
                "example": "startsAt must be a future date and time",
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
      "Event": {
        "properties": {
          "amountMinor": {
            "example": 500000,
            "format": "int64",
            "type": "integer"
          },
          "assignedCrew": {
            "$ref": "#/components/schemas/EventCrew"
          },
          "assignedCrewId": {
            "example": "018f2c1e-6f2a-7c3b-9d4e-2b8a1c5f7e90",
            "nullable": true,
            "type": "string"
          },
          "availableTickets": {
            "example": 99,
            "type": "integer"
          },
          "cancelledAt": {
            "example": null,
            "format": "date-time",
            "nullable": true,
            "type": "string"
          },
          "createdAt": {
            "example": "2026-09-23T17:50:00Z",
            "format": "date-time",
            "type": "string"
          },
          "durationSeconds": {
            "example": 3600,
            "type": "integer"
          },
          "endsAt": {
            "example": "2026-12-01T19:00:00Z",
            "format": "date-time",
            "type": "string"
          },
          "hostId": {
            "example": "018f2c1e-1111-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          },
          "id": {
            "example": "018f2c1e-9a11-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          },
          "name": {
            "example": "Live Show",
            "type": "string"
          },
          "startsAt": {
            "example": "2026-12-01T18:00:00Z",
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "enum": [
              "upcoming",
              "live",
              "ended",
              "cancelled"
            ],
            "example": "upcoming",
            "type": "string"
          },
          "totalTickets": {
            "example": 100,
            "type": "integer"
          },
          "updatedAt": {
            "example": "2026-09-23T17:50:00Z",
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "hostId",
          "name",
          "amountMinor",
          "durationSeconds",
          "status",
          "totalTickets",
          "availableTickets",
          "startsAt",
          "endsAt",
          "createdAt",
          "updatedAt"
        ],
        "type": "object"
      },
      "EventCrew": {
        "properties": {
          "accountId": {
            "example": "018f2c1e-6f2a-7c3b-9d4e-2b8a1c5f7e90",
            "type": "string"
          },
          "availability": {
            "enum": [
              "available",
              "unavailable"
            ],
            "example": "available",
            "type": "string"
          },
          "name": {
            "example": "Crew One",
            "type": "string"
          }
        },
        "required": [
          "accountId",
          "name",
          "availability"
        ],
        "type": "object"
      },
      "EventPage": {
        "properties": {
          "events": {
            "items": {
              "$ref": "#/components/schemas/Event"
            },
            "type": "array"
          },
          "nextCursor": {
            "example": "eyJzIjoiMjAyNi0xMi0wMVQxODowMDowMFoiLCJpIjoiMDE4ZjJjMWUtOWExMS03YzNiLTlkNGUtMmI4YTFjNWY3ZTkwIn0",
            "type": "string"
          }
        },
        "required": [
          "events",
          "nextCursor"
        ],
        "type": "object"
      },
...
      "UpdateEventRequest": {
        "properties": {
          "amountMinor": {
            "format": "int64",
            "type": "integer"
          },
          "assignedCrewId": {},
          "durationSeconds": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "startsAt": {
            "format": "date-time",
            "type": "string"
          },
          "status": {
            "type": "string"
          },
          "totalTickets": {
            "format": "int64",
            "type": "integer"
          }
        },
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "description": "Opaque session token",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Backend for paid live events: event lifecycle, ticket purchase, mocked payment, refunds, and payouts.",
    "title": "LIV DOT API",
    "version": "1.1.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/api/events": {
      "get": {
        "operationId": "listEvents",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "description": "success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "rate limited"
          }
        },
        "summary": "List events with filters",
        "tags": [
          "events"
        ]
      },
      "post": {
        "operationId": "createEvent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            },
            "description": "success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "forbidden"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "rate limited"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Create an event",
        "tags": [
          "events"
        ],
        "x-roles": [
          "host"
        ]
      }
    },
    "/api/events/{id}": {
      "get": {
        "operationId": "getEvent",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EventPage"
                }
              }
            },
            "description": "success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "rate limited"
          }
        },
        "summary": "Get an event and its details",
        "tags": [
          "events"
        ]
      },
      "patch": {
        "operationId": "updateEvent",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEventRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Event"
                }
              }
            },
            "description": "success"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "invalid request"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "unauthenticated"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "forbidden"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "not found"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "conflict"
          },
          "429": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            },
            "description": "rate limited"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Update or cancel an event",
        "tags": [
          "events"
        ],
        "x-roles": [
          "host"
        ]
      }
    },
...
  },
  "servers": [
    {
      "description": "local",
      "url": "http://localhost:8080"
    },
    {
      "description": "same origin",
      "url": "/"
    }
  ]
}